switch everything to D_VEC_XD to start moving to float

+++ src/misc_types.h
@@ -15,8 +15,8 @@ typedef std::uint16_t u_int16_t;
 typedef std::uint32_t u_int32_t;
 typedef std::uint64_t u_int64_t;
 
-typedef struct std::vector<double>                    F_VEC_1D;
-typedef struct std::vector<std::vector<double>>       F_VEC_2D;
+typedef struct std::vector<float>                    F_VEC_1D;
+typedef struct std::vector<std::vector<float>>       F_VEC_2D;
 
 typedef struct std::vector<double>                   D_VEC_1D;
 typedef struct std::vector<std::vector<double>>      D_VEC_2D;
